Tests: experimental in-memory ODB backend
Rather than copying a sandbox repo for every test, which is
expensive on win32, this adds an alternate method which
opens the sandbox in-place, and records ODB writes to an
in-memory hash table.

In cases where it's safe to do so, this speeds up execution immensely.
Running `libgit2_clar.exe -s68` has a visible stutter for each test, but
`-s71 -s72` runs instantaneously.

If configuration, refs, and attributes could be shimmed in a similar
way, this approach could be used for all of the tests, and make running
them bearable on Windows.